Fit and plot CFA simultaneously

Print and save CAF fit, as well as plot CFA loadings, simultaneously.

Usage

cfa_fit_plot(
  model,
  data,
  covs = FALSE,
  remove.items = "",
  file.name,
  estimator = "MLR"
)

Arguments

model

CFA model to fit.

data

Data set on which to fit the CFA model.

covs

Logical, whether to include covariances on the lavaan plot.

remove.items

Optional, if one wants to remove items from the CFA model without having to redefine it completely again.

file.name

Optional, if one wants something different than the default file name. Defaults to saving to the "/model" subfolder of the working directory. If it doesn't exist, it creates it.

estimator

What estimator to use for the CFA.

Examples

(latent <- list(visual = paste0("x", 1:3),
                textual = paste0("x", 4:6),
                speed = paste0("x", 7:9)))
#> $visual
#> [1] "x1" "x2" "x3"
#> 
#> $textual
#> [1] "x4" "x5" "x6"
#> 
#> $speed
#> [1] "x7" "x8" "x9"
#> 

HS.model <- write_lavaan(latent = latent)
cat(HS.model)
#> ##################################################
#> # [---------------Latent variables---------------]
#> 
#> visual =~ x1 + x2 + x3
#> textual =~ x4 + x5 + x6
#> speed =~ x7 + x8 + x9
#> 

library(lavaan)
#> This is lavaan 0.6-12
#> lavaan is FREE software! Please report any bugs.
fit <- cfa_fit_plot(HS.model, HolzingerSwineford1939)
#> lavaan 0.6-12 ended normally after 42 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        30
#> 
#>   Number of observations                           301
#>   Number of missing patterns                         1
#> 
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                85.306      87.132
#>   Degrees of freedom                                24          24
#>   P-value (Chi-square)                           0.000       0.000
#>   Scaling correction factor                                  0.979
#>     Yuan-Bentler correction (Mplus variant)                       
#> 
#> Model Test Baseline Model:
#> 
#>   Test statistic                               918.852     880.082
#>   Degrees of freedom                                36          36
#>   P-value                                        0.000       0.000
#>   Scaling correction factor                                  1.044
#> 
#> User Model versus Baseline Model:
#> 
#>   Comparative Fit Index (CFI)                    0.931       0.925
#>   Tucker-Lewis Index (TLI)                       0.896       0.888
#>                                                                   
#>   Robust Comparative Fit Index (CFI)                         0.930
#>   Robust Tucker-Lewis Index (TLI)                            0.895
#> 
#> Loglikelihood and Information Criteria:
#> 
#>   Loglikelihood user model (H0)              -3737.745   -3737.745
#>   Scaling correction factor                                  1.093
#>       for the MLR correction                                      
#>   Loglikelihood unrestricted model (H1)      -3695.092   -3695.092
#>   Scaling correction factor                                  1.043
#>       for the MLR correction                                      
#>                                                                   
#>   Akaike (AIC)                                7535.490    7535.490
#>   Bayesian (BIC)                              7646.703    7646.703
#>   Sample-size adjusted Bayesian (BIC)         7551.560    7551.560
#> 
#> Root Mean Square Error of Approximation:
#> 
#>   RMSEA                                          0.092       0.093
#>   90 Percent confidence interval - lower         0.071       0.073
#>   90 Percent confidence interval - upper         0.114       0.115
#>   P-value RMSEA <= 0.05                          0.001       0.001
#>                                                                   
#>   Robust RMSEA                                               0.092
#>   90 Percent confidence interval - lower                     0.072
#>   90 Percent confidence interval - upper                     0.114
#> 
#> Standardized Root Mean Square Residual:
#> 
#>   SRMR                                           0.060       0.060
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Sandwich
#>   Information bread                           Observed
#>   Observed information based on                Hessian
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)   Std.lv  Std.all
#>   visual =~                                                             
#>     x1                0.900    0.100    8.973    0.000    0.900    0.772
#>     x2                0.498    0.088    5.681    0.000    0.498    0.424
#>     x3                0.656    0.080    8.151    0.000    0.656    0.581
#>   textual =~                                                            
#>     x4                0.990    0.061   16.150    0.000    0.990    0.852
#>     x5                1.102    0.055   20.146    0.000    1.102    0.855
#>     x6                0.917    0.058   15.767    0.000    0.917    0.838
#>   speed =~                                                              
#>     x7                0.619    0.086    7.193    0.000    0.619    0.570
#>     x8                0.731    0.093    7.875    0.000    0.731    0.723
#>     x9                0.670    0.099    6.761    0.000    0.670    0.665
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)   Std.lv  Std.all
#>   visual ~~                                                             
#>     textual           0.459    0.073    6.258    0.000    0.459    0.459
#>     speed             0.471    0.119    3.954    0.000    0.471    0.471
#>   textual ~~                                                            
#>     speed             0.283    0.085    3.311    0.001    0.283    0.283
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)   Std.lv  Std.all
#>    .x1                4.936    0.067   73.473    0.000    4.936    4.235
#>    .x2                6.088    0.068   89.855    0.000    6.088    5.179
#>    .x3                2.250    0.065   34.579    0.000    2.250    1.993
#>    .x4                3.061    0.067   45.694    0.000    3.061    2.634
#>    .x5                4.341    0.074   58.452    0.000    4.341    3.369
#>    .x6                2.186    0.063   34.667    0.000    2.186    1.998
#>    .x7                4.186    0.063   66.766    0.000    4.186    3.848
#>    .x8                5.527    0.058   94.854    0.000    5.527    5.467
#>    .x9                5.374    0.058   92.546    0.000    5.374    5.334
#>     visual            0.000                               0.000    0.000
#>     textual           0.000                               0.000    0.000
#>     speed             0.000                               0.000    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)   Std.lv  Std.all
#>    .x1                0.549    0.156    3.509    0.000    0.549    0.404
#>    .x2                1.134    0.112   10.135    0.000    1.134    0.821
#>    .x3                0.844    0.100    8.419    0.000    0.844    0.662
#>    .x4                0.371    0.050    7.382    0.000    0.371    0.275
#>    .x5                0.446    0.057    7.870    0.000    0.446    0.269
#>    .x6                0.356    0.047    7.658    0.000    0.356    0.298
#>    .x7                0.799    0.097    8.222    0.000    0.799    0.676
#>    .x8                0.488    0.120    4.080    0.000    0.488    0.477
#>    .x9                0.566    0.119    4.768    0.000    0.566    0.558
#>     visual            1.000                               1.000    1.000
#>     textual           1.000                               1.000    1.000
#>     speed             1.000                               1.000    1.000
#> 
#> R-Square:
#>                    Estimate
#>     x1                0.596
#>     x2                0.179
#>     x3                0.338
#>     x4                0.725
#>     x5                0.731
#>     x6                0.702
#>     x7                0.324
#>     x8                0.523
#>     x9                0.442
#> 
#> Error: Package "DiagrammeRsvg" needed for this function to work. Please install it.